+++ CHANGELOG.md
@@ -1,11 +1,21 @@
 # Change log
+## v0.2.0 Add AWS template
+- This is the first AWS template. 
+- There is a hub-spoke network architecture in this template, core VPC, CDP VPC and transit gateway.
+- Egress from core VPC.
+- CDP prerequisite resources are created in this template. Plus cross account role and public key.
+- Future features:
+  - Egress controlled by AWS network firewall
+  - Custom private DNS server.
+  - RAZ support.
+  - KMS support.
 
-## v0.1.5 core subnet route egress from firewall
+## v0.1.5 azure core subnet route egress from firewall
 - Add route table to core subnet in hub vnet to control core subnet egress from firewall.
   - Add firewall rule to allow egress from core subnet
 - ignore access policy changes in KeyVault
 
-## v0.1.5 more fw rules
+## v0.1.5 more azure fw rules
 - Add a few firewall rules for microsft tools. They are required to install Azure CLI 
 
 ## v0.1.4 azure template